moved more property section scripts
