--- Assets/Scripts/Layers/Properties/PropertyPanels/ILayerCredentialInterface.cs
@@ -1,11 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-namespace Netherlands3D.Twin.Layers.Properties
-{
-    public interface ILayerCredentialInterface
-    {
-        public LayerCredentialsHandler Handler { get; set; }
-    }
-}
