+++ Assets/_Application/Layers/LayerTypes/CartesianTiles/Properties.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 398d1e00bedbf41798b9333f2586db6f
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
