+++ Assets/_Application/Layers/LayerTypes/CartesianTiles/Properties/CartesianTileSubObjectColorPropertyData.cs
@@ -1,11 +1,11 @@
 using System;
 using System.Collections.Generic;
 using System.Runtime.Serialization;
+using Netherlands3D.Twin.Layers.Properties;
 using Newtonsoft.Json;
-using UnityEngine;
 using UnityEngine.Events;
 
-namespace Netherlands3D.Twin.Layers.Properties
+namespace Netherlands3D.Twin.Layers.LayerTypes.CartesianTiles.Properties
 {
     [DataContract(Namespace = "https://netherlands3d.eu/schemas/projects/layers/properties", Name = "CartesianTileSubObjectColor")]
     public class CartesianTileSubObjectColorPropertyData : LayerPropertyData, ILayerPropertyDataWithAssets
