+++ Assets/_Application/Layers/LayerTypes/Credentials/Properties.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c967d31447e744a66ad20a47f0defe93
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
