+++ Assets/_Application/Layers/LayerTypes/Credentials/Properties/CredentialsInputPropertySection.cs
@@ -1,8 +1,7 @@
-using System;
 using TMPro;
 using UnityEngine;
 
-namespace Netherlands3D.Twin.Layers.Properties
+namespace Netherlands3D.Twin.Layers.LayerTypes.Credentials.Properties
 {
     public class CredentialsInputPropertySection : MonoBehaviour, ILayerCredentialInterface
     {
