+++ Assets/_Application/Layers/LayerTypes/Credentials/Properties/ILayerCredentialInterface.cs
@@ -1,0 +1,7 @@
+namespace Netherlands3D.Twin.Layers.LayerTypes.Credentials.Properties
+{
+    public interface ILayerCredentialInterface
+    {
+        public LayerCredentialsHandler Handler { get; set; }
+    }
+}
