+++ Assets/_Application/Layers/LayerTypes/Polygons/Properties/LinePropertySection.cs
@@ -1,7 +1,7 @@
 using UnityEngine;
 using UnityEngine.UI;
 
-namespace Netherlands3D.Twin.Layers.Properties
+namespace Netherlands3D.Twin.Layers.LayerTypes.Polygons.Properties
 {
     public class LinePropertySection : MonoBehaviour
     {
