+++ Assets/_Application/Layers/LayerTypes/Polygons/Properties/LinePropertySectionInstantiator.cs
@@ -1,6 +1,7 @@
+using Netherlands3D.Twin.Layers.Properties;
 using UnityEngine;
 
-namespace Netherlands3D.Twin.Layers.Properties
+namespace Netherlands3D.Twin.Layers.LayerTypes.Polygons.Properties
 {
     public class LinePropertySectionInstantiator : MonoBehaviour, IPropertySectionInstantiator
     {
