+++ Assets/_Application/Layers/Properties/ILayerPropertyDataWithAssets.cs
