+++ Assets/_Application/Layers/Properties/ILayerWithPropertyData.cs
