+++ Assets/_Functionalities/OGC3DTiles/Scripts/Tile3DLayerGameObject.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using Netherlands3D.Tiles3D;
 using Netherlands3D.Twin.Layers;
 using Netherlands3D.Twin.Layers.LayerTypes;
+using Netherlands3D.Twin.Layers.LayerTypes.Credentials;
 using Netherlands3D.Twin.Layers.Properties;
 using UnityEngine;
 using UnityEngine.Events;
