+++ Assets/_Functionalities/Wms/Scripts/WMSLayerGameObject.cs
@@ -1,7 +1,7 @@
 using Netherlands3D.Twin.Layers.Properties;
 using System.Collections.Generic;
 using System.Linq;
-using Netherlands3D.Twin.Layers.LayerTypes;
+using Netherlands3D.Twin.Layers.LayerTypes.CartesianTiles;
 using UnityEngine;
 
 namespace Netherlands3D.Functionalities.Wms
